Add README file; modify city input prompt in main.py
This commit introduces a README file to provide instructions about the use, setup and sample outputs of the weather app. The documentation will help new users or contributors to get familiar with the project. In addition to this, a change in main.py is made to modify the input prompt for the city parameter, making it optional instead of a mandatory input field. This is done to improve the user experience and to set a default fallback mechanism to the weather query for 'Pune' if no city is provided by the user.